Change window to fixed size and retain mouse position on fullscreen toggle
Previously, when toggling fullscreen, the mouse would be set to strange locations, but with this change the mouse position gets reset to its original location.

As for the fixed size change, the window looked too big on 2k and 4k monitors and I now think keeping the window at a constant size looks the best.